refactor: cart price util
